chore(telemetry): Make sure telemetry is sent where expected
Might revisit this again soon.
