+++ src/documentation/commands.ts
@@ -2,7 +2,7 @@ import * as vscode from 'vscode';
 import { DeltaFunctionInfo, DeltaIssue } from '../code-health-monitor/tree-model';
 import { CodeSceneTabPanel } from '../codescene-tab/webview-panel';
 import { FnToRefactor } from '../refactoring/capabilities';
-import { registerCommandWithTelemetry } from '../utils';
+import { registerCommandWithTelemetry } from '../telemetry';
 
 export function register(context: vscode.ExtensionContext) {
   context.subscriptions.push(
