+++ src/utils.ts
@@ -50,31 +50,3 @@ export async function getLogoUrl(extensionPath: string): Promise<string> {
 export function rangeStr(range: Range) {
   return `[${range.start.line + 1}:${range.start.character}→${range.end.line + 1}:${range.end.character}]`;
 }
-
-/**
- * Register a command with this function to automatically log telemetry when executed.
- * Essentially wraps vscode.commands.registerCommand, adding Telemetry.logUsage with optional
- * eventData.
- */
-export function registerCommandWithTelemetry({
-  commandId,
-  handler,
-  thisArg,
-  logArgs,
-}: {
-  commandId: string;
-  handler: (...args: any[]) => any;
-  thisArg?: any;
-  logArgs?: (...args: any[]) => any;
-}): vscode.Disposable {
-  const wrappedHandler = (...args: any[]) => {
-    const eventName = `command/${commandId}`;
-    let eventData;
-    if (isDefined(logArgs)) {
-      eventData = logArgs(...args);
-    }
-    Telemetry.instance.logUsage(eventName, eventData);
-    return handler.apply(thisArg, args);
-  };
-  return vscode.commands.registerCommand(commandId, wrappedHandler, thisArg);
-}
